+++ packages/lib/src/api.ts
@@ -1,0 +1,11 @@
+export type PluginWebApi = {
+	name: string;
+	version: string;
+	enabled: boolean;
+	loaded: boolean;
+	web: {
+		main: string;
+		error?: string;
+	};
+	requirePath: string;
+}
