+++ packages/web_ui/src/components/PluginsPage.tsx
@@ -2,7 +2,7 @@ import React, { useContext, useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { Table } from "antd";
 
-import type { PluginWebApi } from "@clusterio/controller/src/routes";
+import type { PluginWebApi } from "@clusterio/lib";
 
 import notify from "../util/notify";
 import ControlContext from "./ControlContext";
